add automatic promethues metrics for roundtrip db query times
This change adds a prometheus Histogram to each database connection. The
promauto.Histogram will automatically register with the default
prometheus client, if one exists. This change also adds a set of methods
to the the coresql.DB that override the Exec/Query methods of the
underlying sql.DB. These methods record, to the prometheus histogram, the
duration of the Exec/Query call to the underlying sql.DB.

If the application uses the core metrics server from
LUSHDigital/core/workers/metricsrv, the database histogram will be available
via that metrics server's endpoint without any further changes to your
code.